代码提交

+++ framework/dao/src/main/java/com/ruban/framework/dao/IRubanMapper.java
@@ -52,7 +52,7 @@ public interface IRubanMapper {
      * 
      * @param obj
      */
-    <T> Long insert(T obj);
+    <T> int insert(T obj);
 
     /**
      * 更新数据
